// ui-angular/projects/viewmodel/src/lib/core/api/policy/model/viewModelDataSimple.ts
/**
 * Navigation API
 * Navigation API
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: info@it4all.hu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NavigationTarget } from './navigationTarget';


/**
 * Represent a viewmodel\'s data.
 */
export interface ViewModelDataSimple { 
    uuid: string;
    path?: string;
    navigationTarget?: NavigationTarget;
    children: { [key: string]: ViewModelDataSimple; };
}