// ui-angular/projects/viewmodel/src/lib/core/api/policy/model/message.ts
/**
 * Navigation API
 * Navigation API
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: info@it4all.hu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MessageResult } from './messageResult';
import { MessageType } from './messageType';


export interface Message { 
    /**
     * Unique identifier of message. Will be generated if left empty.
     */
    uuid: string;
    /**
     * Identifier of viewModel to refresh after message is answered.
     */
    viewModelUuid?: string;
    type: MessageType;
    header?: string;
    text?: string;
    possibleResults: Array<MessageResult>;
    selectResult?: MessageResult;
}