// ui-angular/projects/viewmodel/src/lib/core/api/policy/model/commandData.ts
/**
 * Navigation API
 * Navigation API
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: info@it4all.hu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CommandData { 
    model?: object;
    commandPath?: string;
    commandCode?: string;
    params?: Array<object>;
}